requirements-ocr.txt
numpy==1.26.4
opencv-python-headless
pillow
pdf2image
paddleocr
paddlepaddle
torch==2.2.2
torchvision==0.17.2
transformers==4.38.2
sentencepiece
easyocr
timm
pytorch-lightning
nltk
lmdb
